mudando o code MOTOR_CONTROL agora funciona RPM

+++ CODE/CONTROL_MOTOR/CONTROL_MOTOR/setDuty_Motor_L.ino
@@ -1,6 +1,6 @@
 void setDuty_Motor_L(float value) {
 
-  int duty;
+  byte duty;
 
   value = value / 100;
